Refactor `uno_helpers.py` by splitting concerns into focused modules
Resolves the Kitchen-sink utility class issue #5 by entirely removing `uno_helpers.py` and modularizing its internal functions to match their domain contexts:
- Moved environment & context helpers (`get_desktop`, `get_active_document`, `get_package_info`, `get_extension_url`, `get_extension_path`) to `uno_context.py`.
- Moved document state checkers and property modifiers (`is_writer`, `is_calc`, `is_draw`, `get_document_property`, `set_document_property`) to `document.py`.
- Moved dialog controls and listeners (`get_optional`, `is_checkbox_control`, `get_checkbox_state`, `set_checkbox_state`, `TabListener`) to `dialogs.py`.

Globally updated imports across all modules, logic files, and test files to reflect the new structure. Updated `AGENTS.md` context file as well.
