+++ plugin/tests/uno/core_tests.py
@@ -1,5 +1,5 @@
 from plugin.framework.document import DocumentCache
-from plugin.framework.uno_helpers import get_desktop
+from plugin.framework.uno_context import get_desktop
 from plugin.testing_runner import setup, teardown, native_test
