+++ plugin/tests/uno/test_calc.py
@@ -15,7 +15,7 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-from plugin.framework.uno_helpers import get_desktop
+from plugin.framework.uno_context import get_desktop
 from plugin.testing_runner import setup, teardown, native_test
