+++ plugin/tests/uno/test_chatbot_integration.kcc
@@ -1,4 +1,4 @@
-from plugin.framework.uno_helpers import get_desktop
+from plugin.framework.uno_context import get_desktop
 from plugin.testing_runner import setup, teardown, native_test
 from plugin.modules.chatbot.panel import ChatSession, SendButtonListener
