+++ plugin/tests/uno/test_writer_navigation.py
@@ -102,7 +102,7 @@ class TestWriterNavigation(unittest.TestCase):
 
 try:
     from plugin.testing_runner import setup, teardown, native_test
-    from plugin.framework.uno_helpers import get_desktop
+    from plugin.framework.uno_context import get_desktop
     from plugin.modules.writer.navigation import NavigateHeading, GetSurroundings  # ToolBaseDummy: not exposed
 except ImportError:
     setup, teardown, native_test = (lambda f: f), (lambda f: f), (lambda f: f)
