+++ plugin/tests/uno/test_writer_search.py
@@ -6,7 +6,7 @@ except ImportError:
 
 try:
     from plugin.testing_runner import setup, teardown, native_test
-    from plugin.framework.uno_helpers import get_desktop
+    from plugin.framework.uno_context import get_desktop
     from plugin.modules.writer.search import SearchInDocument, GetIndexStats
 except ImportError:
     setup, teardown, native_test = (lambda f: f), (lambda f: f), (lambda f: f)
